# pytorrent/openapi/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "error": {
            "type": "string"
          },
          "ok": {
            "enum": [
              false
            ],
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "error"
        ],
        "type": "object"
      },
      "ApiOk": {
        "properties": {
          "ok": {
            "enum": [
              true
            ],
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
...
      "AppStatusResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiOk"
          },
          {
            "properties": {
              "status": {
                "$ref": "#/components/schemas/AppStatus"
              }
            },
            "required": [
              "status"
            ],
            "type": "object"
          }
        ]
      },
...
      "CleanupResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiOk"
          },
          {
            "properties": {
              "cleanup": {
                "$ref": "#/components/schemas/CleanupSummary"
              },
              "deleted": {
                "oneOf": [
                  {
                    "type": "integer"
                  },
                  {
                    "additionalProperties": true,
                    "type": "object"
                  }
                ]
              }
            },
            "type": "object"
          }
        ]
      },
...
      "Job": {
        "additionalProperties": true,
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "description": "Optional per-user API token. Tokens are accepted only when PYTORRENT_AUTH_ENABLE=1. When auth is disabled, API endpoints do not require credentials.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "bearerApiToken": {
        "description": "Optional per-user API token in Authorization: Bearer <token>. Tokens are accepted only when PYTORRENT_AUTH_ENABLE=1.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Optional Flask session cookie. Required only when PYTORRENT_AUTH_ENABLE=1 and no API token is used.",
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "pyTorrent API",
    "version": "0.0.1"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/app/status": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppStatusResponse"
                }
              }
            },
            "description": "Application status"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "pyTorrent application status"
      }
    },
...
    "/api/cleanup/all": {
      "post": {
        "description": "Clears finished job logs, Smart Queue history, active-profile operation logs, Planner action history and automation history. Pending/running jobs, saved rules, settings and torrents are preserved.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Cleanup result"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear all cleanup-supported history"
      }
    },
    "/api/cleanup/automations": {
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Cleanup result"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear automation execution history"
      }
    },
    "/api/cleanup/cache": {
      "post": {
        "description": "Clears active profile runtime and database-backed caches without deleting torrents, settings, rules or logs.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Cleanup result"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No active profile"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear active profile cache"
      }
    },
    "/api/cleanup/jobs": {
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Cleanup result"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear finished job history"
      }
    },
    "/api/cleanup/operation-logs": {
      "post": {
        "description": "Clears active-profile operation logs. Torrents, jobs, settings and rules are preserved.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Cleanup result"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear operation logs"
      }
    },
    "/api/cleanup/planner": {
      "post": {
        "description": "Deletes Download Planner action history for the active profile. Saved Planner settings are preserved.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Cleanup result"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No active profile"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear Planner action history"
      }
    },
    "/api/cleanup/smart-queue": {
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Cleanup result"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear Smart Queue history"
      }
    },
    "/api/cleanup/summary": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            },
            "description": "Cleanup summary"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Cleanup summary"
      }
    },
...
    "/api/jobs": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 50,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "jobs": {
                          "items": {
                            "$ref": "#/components/schemas/Job"
                          },
                          "type": "array"
                        },
                        "total": {
                          "type": "integer"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "List job queue history"
      }
    },
    "/api/jobs/clear": {
      "post": {
        "description": "Deletes jobs that are not pending or running.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "deleted": {
                          "type": "integer"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear finished job history"
      }
    },
    "/api/jobs/{job_id}/cancel": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOk"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Cancel pending or failed job"
      }
    },
    "/api/jobs/{job_id}/force": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOk"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Force pending job to run next"
      }
    },
    "/api/jobs/{job_id}/retry": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOk"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Retry failed or cancelled job"
      }
    },
...
    "/api/path/browse": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PathBrowseResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Browse server directories"
      }
    },
    "/api/path/default": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PathDefaultResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get active default download path"
      }
    },
...
    "/api/port-check": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortCheckStatus"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Read cached incoming port check status"
      },
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortCheckStatus"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Run incoming port check immediately, bypassing cache"
      }
    },
...
    "/api/rtorrent-config/reset": {
      "post": {
        "description": "Clear pyTorrent-saved rTorrent config overrides and reload live rTorrent values.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RtorrentConfigResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Reset startup rTorrent config overrides"
      }
    },
...
    "/api/system/disk": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemDiskResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Disk usage for monitored paths"
      }
    },
...
  }
}